# storage/api-docs/api-spec.yaml
openapi: 3.0.0
info:
  title: 'MegaSorpresa API'
  description: 'API Backend para el e-commerce de juguetes MegaSorpresa. Sirve a clientes Web (SPA), Android e iOS.'
  contact:
    email: api@megasorpresa.com
  version: 1.0.0
servers:
  -
    url: /
    description: 'API Server'
paths:
...
  /api/catalog/products:
    get:
      tags:
        - Catálogo
      summary: 'Listar productos del catálogo'
      description: 'Obtiene un listado paginado de productos con filtros opcionales por ciudad, categoría y búsqueda. Los productos se filtran por disponibilidad en la ciudad especificada.'
      operationId: 16d32315eaac159466bce226dcaad470
      parameters:
        -
          name: city_id
          in: query
          description: 'ID de la ciudad para filtrar productos por disponibilidad'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: category_id
          in: query
          description: 'ID de la categoría para filtrar productos'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: search
          in: query
          description: 'Término de búsqueda para filtrar productos por nombre o descripción'
          required: false
          schema:
            type: string
            example: muñeca
        -
          name: per_page
          in: query
          description: 'Número de productos por página (paginación)'
          required: false
          schema:
            type: integer
            default: 15
            example: 15
      responses:
        '200':
          description: 'Listado de productos obtenido exitosamente'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Product' } }
                type: object
  /api/catalog/products/featured:
    get:
      tags:
        - Catálogo
      summary: 'Listar productos destacados'
      description: 'Obtiene un listado de productos destacados para una ciudad específica. Útil para mostrar productos en la página principal.'
      operationId: ef7353267102607463681d1d87a9b85f
      parameters:
        -
          name: city_id
          in: query
          description: 'ID de la ciudad para obtener productos destacados'
          required: false
          schema:
            type: integer
            default: 1
            example: 1
        -
          name: limit
          in: query
          description: 'Número máximo de productos destacados a retornar'
          required: false
          schema:
            type: integer
            default: 10
            example: 10
      responses:
        '200':
          description: 'Listado de productos destacados obtenido exitosamente'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Product' } }
                type: object
  '/api/catalog/products/{id}':
    get:
      tags:
        - Catálogo
      summary: 'Obtener detalle de un producto'
      description: 'Obtiene la información completa de un producto específico, incluyendo imágenes, categorías y reseñas.'
      operationId: ed75756d0f7ac8d74374f8fff04d19bb
      parameters:
        -
          name: id
          in: path
          description: 'ID del producto'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 'Detalle del producto obtenido exitosamente'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Product' }
                type: object
        '404':
          description: 'Producto no encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /api/reviews:
    get:
      tags:
        - Reseñas
      summary: 'Listar reseñas de un producto'
      description: 'Obtiene todas las reseñas aprobadas de un producto específico. Se requiere el parámetro product_id.'
      operationId: 9cfedb9e60a10482ea2d0017aa8b381d
      parameters:
        -
          name: product_id
          in: query
          description: 'ID del producto para obtener sus reseñas'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 'Listado de reseñas obtenido exitosamente'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { properties: { id: { type: integer, example: 1 }, product_id: { type: integer, example: 1 }, user_id: { type: integer, example: 5 }, rating: { type: integer, maximum: 5, minimum: 1, example: 5 }, comment: { type: string, example: 'Excelente producto, muy recomendado' }, is_approved: { type: boolean, example: true }, created_at: { type: string, format: date-time } }, type: object } }
                type: object
        '422':
          description: 'Falta el parámetro product_id'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Product ID is required' }
                type: object
    post:
      tags:
        - Reseñas
      summary: 'Crear una nueva reseña'
      description: 'Permite a un usuario autenticado crear una reseña para un producto. La reseña requiere aprobación antes de ser visible públicamente.'
      operationId: f9bf17f06508aa2dd6aecbb9771092da
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - product_id
                - rating
              properties:
                product_id:
                  description: 'ID del producto a reseñar'
                  type: integer
                  example: 1
                rating:
                  description: 'Calificación del producto (1-5 estrellas)'
                  type: integer
                  maximum: 5
                  minimum: 1
                  example: 5
                comment:
                  description: 'Comentario de la reseña (opcional)'
                  type: string
                  maxLength: 1000
                  example: 'Excelente producto, muy recomendado'
              type: object
      responses:
        '201':
          description: 'Reseña creada exitosamente'
          content:
            application/json:
              schema:
                properties:
                  data: { properties: { id: { type: integer, example: 1 }, product_id: { type: integer, example: 1 }, rating: { type: integer, example: 5 }, comment: { type: string, example: 'Excelente producto' }, is_approved: { type: boolean, example: false } }, type: object }
                type: object
        '401':
          description: 'No autenticado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '422':
          description: 'Datos de validación inválidos'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      security:
        -
          sanctum: []
  /api/reviews/average:
    get:
      tags:
        - Reseñas
      summary: 'Obtener calificación promedio de un producto'
      description: 'Calcula y retorna la calificación promedio de todas las reseñas aprobadas de un producto específico.'
      operationId: 2ac53ca7c17e82fedb9da722adf0eebc
      parameters:
        -
          name: product_id
          in: query
          description: 'ID del producto para obtener su calificación promedio'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 'Calificación promedio obtenida exitosamente'
          content:
            application/json:
              schema:
                properties:
                  product_id: { type: integer, example: 1 }
                  average_rating: { description: 'Promedio de calificaciones (0-5)', type: number, format: float, example: 4.5 }
                type: object
        '422':
          description: 'Falta el parámetro product_id'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Product ID is required' }
                type: object
...
components:
  schemas:
...
    Product:
      title: Product
      description: 'Modelo de producto'
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Muñeca Barbie'
        slug:
          type: string
          example: muneca-barbie
        sku:
          type: string
          example: TOY-001
        base_price:
          type: number
          format: float
          example: 29.99
        description:
          type: string
          example: 'Muñeca Barbie original con accesorios'
        stock_quantity:
          type: integer
          example: 50
        is_active:
          type: boolean
          example: true
        primary_image:
          type: string
          example: 'https://example.com/images/product.jpg'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
    ValidationError:
      title: 'Validation Error'
      description: 'Error de validación (422)'
      properties:
        message:
          type: string
          example: 'The given data was invalid.'
        errors:
          type: object
          example:
            email:
              - 'The email field is required.'
          additionalProperties:
            type: array
            items:
              type: string
      type: object
    UnauthorizedError:
      title: 'Unauthorized Error'
      description: 'Error de autenticación (401)'
      properties:
        message:
          type: string
          example: Unauthenticated.
      type: object
    NotFoundError:
      title: 'Not Found Error'
      description: 'Recurso no encontrado (404)'
      properties:
        message:
          type: string
          example: 'Resource not found.'
      type: object
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Token de autenticación de Laravel Sanctum para clientes móviles'
      bearerFormat: Token
      scheme: bearer
...